requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
evdev==1.7.1
idna==3.10
joblib==1.4.2
keyboard==0.13.5
Markdown==3.7
markdown2==2.5.0
mss==9.0.2
nltk==3.9.1
numpy==2.1.2
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pydub==0.25.1
Pygments==2.18.0
pynput==1.7.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pytesseract==0.3.13
python-dotenv==1.0.1
python-xlib==0.33
pyttsx3==2.98
regex==2024.9.11
requests==2.32.3
six==1.16.0
tkhtmlview==0.3.1
tqdm==4.66.5
urllib3==2.2.3